Corrected docstring example #1506
Update to previous commit, with example for Object dtype.

Corrected the expected results from the operations to account for Struct in polars and Apache Arrow, rather than Object like that returned in Pandas.

This has been conducted in this way since objects are fundamentally challenging in Polars and Apache Arrow compared to pandas (which is more flexible and generally allows use of objects without strict controls of the other libraries).