Refactored Object docstring example #1506
Refactored docstring example narwhals-dev#1506.

Due to issues with Object type in Apache Arrow, example refactored to use a dummy class, which stored within Pandas and Polars as Object dtype.

This is reflected in the example and test used in the documentation.

Apache Arrow example excluded to avoid confusion on Struct dtype being different.