docs: Add Support section to README
Like this, if people want things prioritised beyond what can be done by volunteers, they have an option to speed things up

Note that I would first need to disclose any contacts with my employer (Quansight Labs - they don't own the project in any way, but per my contract with them, I would need to disclose this). Then, if the request was something that they weren't interested in (e.g. if it was too smale-scale) then perhaps one of our regular contributors would be able to take it on independently